+++ doc/RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 ## 2025-10-13 Update
 * Changing a device fan motor or speed now disables rules on that device for 3 hours
 * Air quality variables can now be used in rules
+* Dropdown on chart now only appears if you specify ?dropdown=1